requirements.txt
fastapi
uvicorn
transformers
torch
pillow
python-multipart
langchain
langchain-core
langchain-huggingface
python-dotenv